+++ test/spec/router_test.js
@@ -45,21 +45,10 @@ describe('Router', function() {
     it('should register a beforeNavigation callback', function() {
       var router, method;
 
-      window.Top = {};
-      window.Top.Index = function() {};
-      window.Top.Index.prototype.mount = function() {};
-      method = sinon.spy();
-
       router = new Cape.Router();
-      router._.setHash = function() {};
-      router.draw(function(m) {
-        m.match('', 'top/index');
-      })
-      router.beforeNavigation(method);
-      router.mount('main');
-      router.start();
+      router.beforeNavigation(function() {})
 
-      expect(method.called).to.be(true);
+      expect(router._.beforeNavigationCallbacks.length).to.be(1);
     })
   })
 
@@ -238,43 +227,6 @@ describe('Router', function() {
 
       expect(method.calledWith('main')).to.be(true);
       expect(router.params.id).to.be('123');
-    })
-
-    it('should skip beforeNavigationCallbacks', function() {
-      var router, method;
-
-      window.TestMessage = function() {};
-      window.TestMessage.prototype.mount = function() {};
-      router = new Cape.Router();
-      router._.setHash = function() {};
-      router.draw(function(m) {
-        m.match('hello', 'test_message');
-      })
-      method = sinon.spy();
-      router.beforeNavigation(method);
-      router.mount('main');
-      router.navigate('hello', true);
-
-      expect(method.called).to.be(false);
-    })
-
-    it('should not mount the route is waiting', function() {
-      var router, method;
-
-      window.TestMessage = function() {};
-      window.TestMessage.prototype.mount = method = sinon.spy();
-      router = new Cape.Router();
-      router._.setHash = function() {};
-      router.draw(function(m) {
-        m.match('hello', 'test_message');
-      })
-      router.beforeNavigation(function() {
-        this.waiting = true;
-      });
-      router.mount('main');
-      router.navigate('hello');
-
-      expect(method.called).to.be(false);
     })
   })
 })